Add validation to frontend for forms and reminders

+++ reminder_web/src/pages/ChangePasswordPage.jsx
@@ -3,6 +3,7 @@ import { Link, useNavigate, useLocation } from "react-router-dom"
 import { HeaderFooter } from "../components/HeaderFooter";
 import { Loading } from "../components/Loading"
 import { api } from "../api/api"
+import { validatePassword } from "../validation/validation";
 
 export const ChangePassword = () => {
     const state = useLocation()
@@ -39,20 +40,35 @@ export const ChangePassword = () => {
         e.preventDefault();
         setLoadingState(true)
 
-        if (formData.newPassword === formData.confirmNewPassword) {
-            const response = await api.changeUserPassword(formData);
-            if (response.status === 200) {
-                sessionStorage.clear();
-                navigate("/", { state: {message: "Please log in with your new password."}})
+        if (validatePassword(formData.newPassword)) {
+            if (formData.newPassword === formData.confirmNewPassword) {
+                const response = await api.changeUserPassword(formData);
+                if (response.status === 200) {
+                    sessionStorage.clear();
+                    navigate("/", { state: {message: "Please log in with your new password."}})
+                } else {
+                    console.log(response)
+                    setLoadingState(false)
+                    setError(response.response.data)
+                }
             } else {
-                console.log(response)
                 setLoadingState(false)
-                setError(response.response.data)
+                setError("New passwords do not match!")
             }
         } else {
             setLoadingState(false)
-            setError("New passwords do not match!")
+            setError(<div>
+                <p>The password needs to be:</p>
+                <ul style={{textAlign: "left"}}>
+                    <li>Between 8 and 32 characters</li>
+                    <li>Contain 1 Uppercase character</li>
+                    <li>Contain 1 Lowercase character</li>
+                    <li>Contain 1 Number</li>
+                    <li>Contain 1 special character</li>
+                </ul>
+            </div>)
         }
+
     }
